+++ Sources/VCore/Views/SwiftUI/Base Button (SwiftUI)/SwiftUIBaseButtonState.swift
@@ -9,7 +9,6 @@ import Foundation
 
 // MARK: - Swift UI Base Button State
 /// Enum that describes state, such as `enabled` or `disabled`.
-@available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 public enum SwiftUIBaseButtonState: Int, CaseIterable {