+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Services and managers
 
 Views
 
+- `SwiftUIBaseButton` now supports `macOS`
 - `FetchDelegatingAsyncImage` is added that allows you to write custom fetch function
 - `possible` is added to `BaseButtonGestureState`, and some `case`s are renamed
 - `HOrVStack` is renamed to `HVStack`